Add project methods for supervisor class
This commit adds basic debugging code to ensure that the 3 project
methods for supervisor works as intended.
Supervisors can now view their projects they created, create a new
project 1 at a time, and modify the title of their project.

This change was made to implement 3 out of the 6 required methods
for implemeting FYPMS for Supervisor class.

No bug request fixed.